Drop multi-root guards and keep createRoot off the native import
createTestRoot imported createRoot from renderer.ts, which loads
GpuixRenderer. That made every native-optional test file crash on import
instead of skipping when the .node is missing.

Move createRoot and flushSync next to the reconciler so tests can load
the host config without dlopen. Keep the public re-exports on renderer.ts.

Also drop the per-attach different-root checks. GPUIX is one window.

Session: ses_fcd3def2bffe7YdmGRE7HPRzQA
