+++ packages/react/src/reconciler/reconciler.ts
@@ -1,4 +1,11 @@
+import React from "react"
+import type { ReactNode } from "react"
 import ReactReconciler from "react-reconciler"
+import type { OpaqueRoot } from "react-reconciler"
+import { ConcurrentRoot } from "react-reconciler/constants.js"
+import { GpuixContext } from "../hooks/use-gpuix.js"
+import type { Container, NativeRenderer } from "../types/host.js"
+import { wrapWithBatching } from "./batch-renderer.js"
 import { hostConfig } from "./host-config.js"
 
 // Cast to any because @types/react-reconciler is out of date with react-reconciler 0.31.0
@@ -11,4 +18,67 @@ try {
   reconciler.injectIntoDevTools()
 } catch {
   // DevTools not available
+}
+
+const _r = reconciler as typeof reconciler & {
+  flushSyncFromReconciler?: typeof reconciler.flushSync
+}
+export const flushSync = _r.flushSyncFromReconciler ?? _r.flushSync
+
+export interface Root {
+  render: (node: ReactNode) => void
+  unmount: () => void
+}
+
+export function createRoot(renderer: NativeRenderer): Root {
+  let container: OpaqueRoot | null = null
+  const batchedRenderer = wrapWithBatching(renderer)
+  const gpuixContainer: Container = {
+    renderer: batchedRenderer,
+  }
+
+  const cleanup = (): void => {
+    if (container) {
+      // Must be sync. A late unmount destroy()s remounted ids and the window goes black.
+      flushSync(() => {
+        reconciler.updateContainer(null, container, null, () => {})
+      })
+      container = null
+    }
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  container = (reconciler.createContainer as any)(
+    gpuixContainer,
+    ConcurrentRoot,
+    null,
+    false,
+    null,
+    "",
+    console.error,
+    console.error,
+    console.error,
+    null
+  )
+
+  return {
+    render: (node): void => {
+      const activeContainer = container
+      if (!activeContainer) {
+        throw new Error("Cannot render an unmounted GPUIX root")
+      }
+      reconciler.updateContainer(
+        React.createElement(
+          GpuixContext.Provider,
+          { value: { renderer: batchedRenderer } },
+          node
+        ),
+        activeContainer,
+        null,
+        () => {}
+      )
+    },
+
+    unmount: cleanup,
+  }
 }
